requirements.txt
black==25.1.0
click==8.2.1
iniconfig==2.1.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
psutil==7.0.0
Pygments==2.19.2
pytest==8.4.1
PyYAML==6.0.2
ruff==0.12.5
schedule==1.2.2
